Fixing an error in sync caused by a variable that was not being set. Also refactored syn and moved it into its own package.
