+++ sdk/src/main/java/com/meniga/sdk/models/sync/operators/MenigaSyncOperations.java
@@ -1,8 +1,8 @@
-package com.meniga.sdk.models.user.operators;
+package com.meniga.sdk.models.sync.operators;
 
 import com.meniga.sdk.helpers.Result;
-import com.meniga.sdk.models.user.MenigaSync;
-import com.meniga.sdk.models.user.MenigaSyncStatus;
+import com.meniga.sdk.models.sync.MenigaSync;
+import com.meniga.sdk.models.sync.MenigaSyncStatus;
 
 /**
  * Copyright 2017 Meniga Iceland Inc.
